fix buildingViewTest and remove star imports

+++ data-jdbc/src/test/groovy/io/micronaut/data/jdbc/oraclexe/jsonview/ApartmentId.java
@@ -6,6 +6,6 @@ import jakarta.persistence.Embeddable;
 @Embeddable
 @Serdeable
 public record ApartmentId (
-    int buildingId,
-    int flatId
+    Long buildingId,
+    Long flatId
 ) {}
